[CVW-025] Refactor, SortSelectionView가 RO기반으로 상태를 변경하도록 수정
